feat: v2.2.7 — 2차 전체 로직 감사 반영 + 사장님 결정 5건 (2026-09-22~23)
6갈래 감사에서 재현된 결함만 고치고, 고친 자리마다 AVR 시뮬레이터 행위 시험을 붙였다
(tools/sim — 제품 코드를 실물 헤더·LTO 로 빌드해 avr-gdb 시뮬레이터에서 setup/loop 실행).
수정 전 코드에서 빨강 → 수정 후 초록을 확인했고, 수정을 하나씩 되돌린 변이 21건이 모두 해당
시험만 빨강이 됐다. 판정·근거는 docs/AUDIT_2026-09-22.md.

감사 결함
- 소독기 RTC 복구: 세척 종료 기록이 0 이면 unixtime 이 돌아 RTC 가 2043-04-29 에 굳던 것
  → 종료가 무효·지난 주기면 "세척 시작 + 설정 세척시간 + 1분"(사장님 09-23)
- 일회성 매니저 거부 경로에서 이동 플래그(mMovable) 미해제 — 08-11 수정의 형제 누락
- 세척·소독 시작 기록 실패에 성공음·알람·host 등록 → Write Error 경고, 재접촉 유도
- 더블터치 2초 가드가 host 시각에만 묶여 guest 튐이 0분 종료 → "마지막으로 시작한 그 태그"
- 인증 실패·HaltA 유실 뒤 올려 둔 태그를 3루프마다 재처리 → Poll 3회 무응답만 이탈
- 서버 덤프에 읽기 실패가 있으면 태그를 지우지 않는다(재스캔이 복구 경로)
- 메뉴 숫자 저장이 앞 2자리만 읽던 것(150→15), LCD 날짜 월말·윤년 미검증(2/30 저장),
  JSON 값이 좁혀진 뒤 클램프(300→44), 메뉴 페이지 넘김 무한 재귀(60회 스택 4.4KB)

사장님 결정
- 액교환일: 시계가 방전 표지(2026-01-01)면 미뤘다가 시계가 맞춰질 때 기록(EEPROM 177)
- 게이트웨이 수신 대기 1초 — 올눈(ALLNuN)은 G2~G5 를 250ms 간격으로 따로 보낸다
  (150ms 면 검사일시가 0 으로 기록됐다. 올눈 게이트웨이 현장 있음)
- G3·G4 필드를 블록 16바이트 전부 저장(올눈 16바이트 조각의 끝 한글이 반쪽으로 남던 것)
- 세척기에도 더블터치 2초 방지(소독기와 같은 규칙)

빌드: RAM 3,013 B(36.8%) · Flash 48,332 B · 우리 코드 경고 0 · 시뮬레이터 시험 9종 91개 초록.
★버전이 올랐으므로 업로드 시 EEPROM 이 완전 초기화된다(설정 재입력 필요).

+++ lib/TraceQ_Arduino/src/TraceQ_Arduino/data/nvm/AlarmOption.hpp
@@ -11,10 +11,11 @@ public:
 
     bool   GetFlag() const;
     void   SetFlag(bool flag);
+    // Set 은 int 로 받아 0~127 로 자른 뒤 1바이트로 저장한다(좁힌 뒤 자르면 200→음수, 300→44).
     int8_t GetTimeSlot1() const;
-    void   SetTimeSlot1(int8_t minute);
+    void   SetTimeSlot1(int minute);
     int8_t GetTimeSlot2() const;
-    void   SetTimeSlot2(int8_t minute);
+    void   SetTimeSlot2(int minute);
 
 protected:
     // 1.0과 동일한 EEPROM 주소 (데이터 호환).
